+++ packages/gov-ui-kit/src/core/components/collapsible/index.ts
@@ -1,0 +1,2 @@
+export { Collapsible } from './collapsible';
+export { type CollapsedSize, type ICollapsibleProps } from './collapsible.api';