+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+-   Implement `Collapsible`, and `CardCollapsible` core components
+
 ## [1.0.26] - 2024-05-06
 
 ### Added